Move shared Velero running functionality to velerorunner pkg
